requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.31.0
python-multipart==0.0.20
jinja2==3.1.6
pydantic==2.11.7
python-chess==1.0.0
